perf: streaming compiler and 64-bit artifact offsets (format 4.0)
The compiler materialized the whole CSV as List<String[]>, held several
derived copies, and produced the artifact as one byte[] addressed by 32-bit
offsets. That meant compile heap 3-4x the table as strings, a hard 2 GB
artifact ceiling, and a PRD maximum shape (20M rows x 170 columns, ~27 GB of
column data) that was neither compilable nor representable.

Compilation now streams end to end. Pass 1 reads the CSV collecting exactly
what encoding needs up front: the string dictionary (bounded by distinct
values, not rows), per-column decimal scales, one priority int per row, and
the row count; evaluation order is a stable ascending argsort of the
priorities. Pass 2 reads the CSV again, encoding each cell to its
order-preserving code as it arrives and appending a small self-describing
record to that column's temporary file in source order. A stitch phase then
writes the artifact file sequentially - header, dictionary, definitions, then
one column at a time: load that column's temporary, index its records,
iterate rows in evaluation order emitting the column's subsections, delete
the temporary. Every section offset is known before the first byte is
written, so nothing is patched. Peak compile heap is the dictionary plus one
int per row plus the largest single column - independent of table size.
DecisionTableSource.openStream() must be re-openable, now documented.

Artifact format 4.0: the header grows to 64 bytes with 64-bit section offsets
and an explicit rule_order_offset; column definitions grow to 24 bytes with a
64-bit data_offset; the rule-order section is a 1-byte order type where 0
(identity - always, since rows are physically pre-sorted) carries no array.
CompiledRuleset is file-backed: the compiler streams to a temp file, writeTo
is a copy, and bytes() materializes only artifacts small enough for one
array. The loader maps the artifact per section - metadata once, then each
column's data slice as its own buffer sized by the next column's offset - so
no mapping spans 2 GB, decoders keep their int-indexed ByteBuffer reads, and
load(compiled) maps the backing file directly instead of copying bytes into
an allocateDirect buffer. Each column's data must stay under 2 GB, enforced
with a clear error at compile time.

Measured at 5M rows x 10 EQ string columns (335 MB CSV): compile 58-85 s
down to 45.4 s, and 34.9 s under a 256 MB heap cap - a cap under which the
old compiler could not run at all, since the table alone was ~1.5 GB of
strings. The 60 s compile target is now met with margin.

OffHeapMemoryTest's on-heap assertion now disables index prewarming so it
measures where column data lands, not the indexes - which deliberately live
in direct buffers regardless of load mode.

Design in ADR-0012; format history in docs/artifact-format.md; measurements
in docs/review/engine-design-review.md section 11. The one remaining Tier 2
item is optional: persisting indexes into the artifact at compile time to
cut the 12-13 s load-time index build.

+++ kisoku-runtime/src/main/java/in/systemhalted/kisoku/runtime/loader/ColumnDefinition.java
@@ -23,7 +23,7 @@ record ColumnDefinition(
     ColumnType type,
     ColumnRole role,
     int flags,
-    int dataOffset,
+    long dataOffset,
     int scale) {
 
   /** Flag indicating the column allows null values. */
