+++ kisoku-runtime/src/test/java/in/systemhalted/kisoku/runtime/loader/BinaryArtifactReaderOffsetTest.java
@@ -52,7 +52,7 @@ class BinaryArtifactReaderOffsetTest {
     BinaryArtifactReader reader = compileMultiColumnTable(tempDir);
 
     // dataOffset is relative to the rule-data section base, so the first column starts at 0.
-    assertEquals(0, reader.columns().get(0).dataOffset());
+    assertEquals(0L, reader.columns().get(0).dataOffset());
   }
 
   @Test
@@ -62,9 +62,9 @@ class BinaryArtifactReaderOffsetTest {
     var columns = reader.columns();
     assertTrue(columns.size() >= 2, "need multiple columns to verify offsets");
 
-    int previous = -1;
+    long previous = -1;
     for (var column : columns) {
-      int offset = column.dataOffset();
+      long offset = column.dataOffset();
       assertTrue(
           offset > previous,
           "Column "
